[NPM] Added Faucet_4.sol, Faucet_5.sol, and npm packages
 * Faucet_5_abi.json is a composition of all the file ABIs, not a real JSON

+++ mastering_ethereum/solidity/Faucet.sol
@@ -3,8 +3,8 @@ pragma solidity ^0.6.0;
 
 contract Faucet {
     receive() external payable {}
-    
-    function withdraw(uint withdraw_amount) public {
+
+    function withdraw(uint256 withdraw_amount) public {
         require(withdraw_amount <= 100000000000000000);
         msg.sender.transfer(withdraw_amount);
     }
